Code refactoring (created class Tasks)
